Fixed highlighting tab completion without zsh-syntax-highlighting
When pressing tab on an incomplete command like 'l' or 'c' the commandline
turns and stays green so even the output is colored green or the
highlight of the suggestions swaps colors when tabbing through.
This happens only if the zsh-syntax-highlighting plugin is not enabled.
I fixed this by adding the missing %f in prompt_char().